Add some rails convention into dotnet (#161)
* Add some rails convention into dotnet
* Add redirect for legacy org details url
