+++ src/ui/Views/Organisation/Index.cshtml
@@ -14,7 +14,7 @@
         {
           <li>
             <h3 class="govuk-heading-m">
-              <a href="@Url.Action("Courses", "Organisation", new {ucasCode = organisation.UcasCode})" class="govuk-link">@organisation.OrganisationName</a>
+              <a href="@Url.Action("Show", "Organisation", new {ucasCode = organisation.UcasCode})" class="govuk-link">@organisation.OrganisationName</a>
               <span class="govuk-body govuk-!-font-weight-regular govuk-!-display-block">@organisation.TotalCourses courses</span>
             </h3>
           </li>
